FIX: [마이페이지] 우리동네 플로깅 & 플로깅 모임 화면에서 프로필
